backend/requirements.txt
Django==3.2.15
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.1
djangorestframework-simplejwt==4.7.2
drf-base64==2.0
django-filter==21.1
